+++ OUDS/Core/Components/Sources/_/ViewModifiers/TypographyModifiers/TypographyModifier.swift
@@ -54,7 +54,7 @@ struct TypographyModifier: ViewModifier {
         isCompactMode ? font.compact : font.regular
     }
 
-    #if canImport(UIKit)
+    #if os(iOS) || os(visionOS)
     /// According to the current `OUDSTheme` and if a custom font is applied or not, returns the suitable `Font`
     private var adaptiveTypography: Font {
         if let family {
@@ -93,7 +93,11 @@ struct TypographyModifier: ViewModifier {
                 .lineSpacing(adaptiveLineHeight)
                 .tracking(adaptiveFont.letterSpacing)
                 .minimumScaleFactor(0.5)
+            #if os(visionOS)
+                .onChange(of: sizeCategory) { _, _ in }
+            #else
                 .onChange(of: sizeCategory) { _ in }
+            #endif
         } else {
             content
                 .font(adaptiveTypography)
@@ -101,6 +105,50 @@ struct TypographyModifier: ViewModifier {
                 .minimumScaleFactor(0.5)
                 .onChange(of: sizeCategory) { _ in }
         }
+    }
+
+    #elseif os(macOS)
+    /// According to the current `OUDSTheme` and if a custom font is applied or not, returns the suitable `Font`
+    private var adaptiveTypography: Font {
+        if let family {
+            // Can be a custom font load from side assets or another custom font available in the OS
+            let composedFontFamily = kApplePostScriptFontNames[orKey: PSFNMK(family, adaptiveFont.weight.weight)]
+            let customFont: Font = .custom(composedFontFamily, size: adaptiveFont.size)
+            return customFont
+        } else {
+            // Apply the system font with weight, responsive to Dynamic Type
+            return .system(size: scaledFontSize, weight: adaptiveFont.weight.weight, design: .default)
+        }
+    }
+
+    /// Computes the line spacing value to apply on the typography.
+    /// Difference between the line height token and line height of computed font to apply depending the the scaled font size.
+    /// The line spacing token cannot be used as is.
+    private var adaptiveLineHeight: CGFloat {
+        let layoutManager = NSLayoutManager()
+        let systemFont = NSFont.systemFont(ofSize: scaledFontSize)
+        let systemLineHeight = layoutManager.defaultLineHeight(for: systemFont)
+        return adaptiveFont.lineHeight - systemLineHeight
+    }
+
+    /// Adjusts the font size dynamically based on the user's accessibility settings
+    /// using UIFontMetrics to scale the font size, ensuring Dynamic Type support
+    private var scaledFontSize: CGFloat {
+        adaptiveFont.size
+    }
+
+    /// Applies to the `Content` the *adaptive font* (i.e. *font family*, *font weight* and *font size*
+    /// depending to the current `MultipleFontCompositeRawTokens`.
+    func body(content: Content) -> some View {
+        // `tracking()` only available for iOS 16+
+        // `minimumScaleFactor()` ensures text remains readable by allowing scaling down
+        // `.onChange(of: sizeCategory) { _ in }` triggers view update when Dynamic Type size changes
+        content
+            .font(adaptiveTypography)
+            .lineSpacing(adaptiveLineHeight)
+            .tracking(adaptiveFont.letterSpacing)
+            .minimumScaleFactor(0.5)
+            .onChange(of: sizeCategory) { _, _ in }
     }
     #else
     /// Applies to the `Content` the *adaptive font* (i.e. *font family*, *font weight* and *font size*
@@ -113,7 +161,7 @@ struct TypographyModifier: ViewModifier {
             content
                 .tracking(adaptiveFont.letterSpacing)
                 .minimumScaleFactor(0.5)
-                .onChange(of: sizeCategory) { _ in }
+                .onChange(of: sizeCategory) { _, _ in }
         } else {
             content
                 .minimumScaleFactor(0.5)