+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for `visionOS` (Orange-OpenSource/ouds-ios#1102) 
+- Support for `macOS` (Orange-OpenSource/ouds-ios#1102) 
 - `input tag` component (Orange-OpenSource/ouds-ios#917)
 - Swift DocC tutorial (Orange-OpenSource/ouds-ios#919)